refactor: remove the dependency on pallas-network in amaru-consensus
