+++ crates/amaru-consensus/Cargo.toml
@@ -23,7 +23,7 @@ hex.workspace = true
 itertools.workspace = true
 opentelemetry.workspace = true
 pallas-crypto.workspace = true
-pallas-network.workspace = true
+pallas-primitives.workspace = true
 proptest = { workspace = true, optional = true }
 rand = { workspace = true, optional = true }
 rayon.workspace = true
@@ -38,13 +38,15 @@ amaru-kernel.workspace = true
 amaru-ouroboros.workspace = true
 amaru-mempool.workspace = true
 amaru-metrics.workspace = true
-amaru-network.workspace = true
 amaru-ouroboros-traits.workspace = true
 amaru-slot-arithmetic.workspace = true
 pure-stage.workspace = true
 
 [dev-dependencies]
+amaru-network.workspace = true
+async-trait.workspace = true
 hex.workspace = true
+pallas-network.workspace = true
 pallas-primitives.workspace = true
 proptest.workspace = true
 rand.workspace = true
