+++ crates/amaru-consensus/README.md
@@ -3,8 +3,7 @@
 ## Chain Diffusion Pipeline
 
 The following graph represents a simplified sequence of _stages_ a block undergoes in
-the consensus module to implement chain synchronisation and diffusion. Each square node is a processing step, and arrows
-are labelled with the type of messages flowing between the various steps.
+the consensus module to implement chain synchronisation and diffusion. Each square node is a processing step, and arrows are labelled with the type of messages flowing between the various steps.
 
 ```mermaid
 graph TD
@@ -47,26 +46,20 @@ graph TD
 
 ### Stages
 
-* [pull](src/consensus/stages/pull_chain_sync_events.rs): connects to upstream peers, running chain sync and block fetch
-  protocols.
-* [receive header](src/consensus/stages/receive_header.rs): this stage is responsible for basic sanity check of _chain
-  sync_
+* [pull](src/consensus/stages/pull.rs): connects to upstream peers, running chain sync and block fetch protocols.
+* [receive header](src/consensus/stages/receive_header.rs): this stage is responsible for basic sanity check of _chain sync_
   messages, deserialising raw headers, and potentially checking whether or not they should be further processed (eg. if
   a header is already known to be invalid, or known to be valid because it's part of our best chain, let's not waste
   time processing it!). The received header is stored at this stage, indexed by its hash.
-* [track_peers](src/consensus/stages/track_peers.rs): this stage keeps track of upstream peers' state and logs a message
-  to notify operators when we are caught up with an upstream peer.
-* [validate header](src/consensus/stages/validate_header.rs): protocol validation of the header, checks the correctness
-  of the
+* [track_peers](src/consensus/stages/track_peers.rs): this stage keeps track of upstream peers' state and logs a message to notify operators when we are caught up with an upstream peer.
+* [validate header](src/consensus/stages/validate_header.rs): protocol validation of the header, checks the correctness of the
   VRF leader election w.r.t relevant stake distribution, and epoch nonce
 * [fetch block](src/consensus/stages/fetch_block.rs): fetch block body corresponding to the new header, if any.
   The block body is stored, indexed by header hash.
   It is stored _before_ validation in order to later
   support [pipelining](https://iohk.io/en/blog/posts/2022/02/01/introducing-pipelining-cardanos-consensus-layer-scaling-solution/).
-* [validate block](src/consensus/effects/ledger_effects.rs): validate the block body against its parent ledger state and
-  apply it
-* [select chain](src/consensus/stages/select_chain.rs): proceed to chain (candidate) selection, possibly changing the
-  current
+* [validate block](src/consensus/effects/ledger_effects.rs): validate the block body against its parent ledger state and apply it
+* [select chain](src/consensus/stages/select_chain.rs): proceed to chain (candidate) selection, possibly changing the current
   best chain,
 * [forward chain](src/consensus/stages/forward_chain.rs): forward newly selected chain to downstream peers (chain
   followers)
@@ -101,20 +94,19 @@ are considered fatal and cause the node to panic.
 
 The consensus chain synchronisation pipeline emits the following spans (at `TRACE` level) when traces are enabled:
 
-* `diffusion.chain_sync`: toplevel trace and entry-point to the `ChainSync` protocol pipeline. Basically handles 2 types
-  of messages: `RollForward <header>` and `Rollback <point>` from each peer the node is connected to
+* `diffusion.chain_sync`: toplevel trace and entry-point to the `ChainSync` protocol pipeline. Basically handles 2 types of messages: `RollForward <header>` and `Rollback <point>` from each peer the node is connected to
 * `chain_sync.receive_header`: sanity checks for headers
-    * `chain_sync.decode_header`: decoding of CBOR content of header
-    * `consensus.store.store_header`: store header on disk in ChainDB
+  * `chain_sync.decode_header`: decoding of CBOR content of header
+  * `consensus.store.store_header`: store header on disk in ChainDB
 * `chain_sync.track_peers`: keep track of whether or not the node is _caught-up_ w.r.t. each of its peer
 * `chain_sync.validate_header`: full validation of received headers
-    * `validate_header.evolve_nonce`: apply nonces evolution logic
-    * `validate_header.validate`: check all headers validity rules (eg. VRF proofs, signatures, consistency...)
+  * `validate_header.evolve_nonce`: apply nonces evolution logic
+  * `validate_header.validate`: check all headers validity rules (eg. VRF proofs, signatures, consistency...)
 * `diffusion.fetch_block`: retrieve block body from peer(s)
-    * `consensus.store.store_block`: store block on disk in ChainDB
+  * `consensus.store.store_block`: store block on disk in ChainDB
 * `chain_sync.validate_block`: full validation of the block body against current tip's ledger state
 * `chain_sync.select_chain`: apply chain selection logic based on the given protocol message
-    * `consensus.store.roll_forward_chain`: store current chain state on disk
+  * `consensus.store.roll_forward_chain`: store current chain state on disk
 * `diffusion.forward_chain`: propagage possible changes in node's best chain to downstream peers
 
 Tracing data can be collected through the OpenTelemetry exported, here is a rendering of such a trace using Jaeger UI:
