+++ crates/amaru-consensus/src/consensus/effects/consensus_effects.rs
@@ -109,14 +109,14 @@ pub mod tests {
     use amaru_kernel::connection::ClientConnectionError;
     use amaru_kernel::is_header::HeaderTip;
     use amaru_kernel::peer::Peer;
-    use amaru_kernel::{Point, PoolId, RawBlock, TxId};
+    use amaru_kernel::{Point, PoolId, RawBlock};
     use amaru_mempool::strategies::InMemoryMempool;
     use amaru_metrics::MetricsEvent;
     use amaru_metrics::ledger::LedgerMetrics;
     use amaru_ouroboros_traits::can_validate_blocks::HeaderValidationError;
     use amaru_ouroboros_traits::in_memory_consensus_store::InMemConsensusStore;
     use amaru_ouroboros_traits::{
-        BlockValidationError, HasStakeDistribution, PoolSummary, TxSubmissionMempool,
+        BlockValidationError, HasStakeDistribution, PoolSummary, TxId, TxSubmissionMempool,
     };
     use amaru_slot_arithmetic::Slot;
     use pure_stage::{BoxFuture, Instant, StageRef};
