+++ crates/amaru-consensus/src/consensus/effects/mempool_effects.rs
@@ -13,10 +13,9 @@
 // limitations under the License.
 
 use amaru_kernel::Tx;
-use amaru_kernel::tx_submission_events::TxId;
 use amaru_ouroboros_traits::{
-    CanValidateTransactions, MempoolSeqNo, TransactionValidationError, TxOrigin, TxRejectReason,
-    TxSubmissionMempool,
+    CanValidateTransactions, MempoolSeqNo, TransactionValidationError, TxId, TxOrigin,
+    TxRejectReason, TxSubmissionMempool,
 };
 use pure_stage::{
     BoxFuture, Effects, ExternalEffect, ExternalEffectAPI, ExternalEffectSync, Resources, SendData,
@@ -289,9 +288,9 @@ impl ExternalEffectSync for LastSeqNo {}
 
 #[cfg(test)]
 mod tests {
-    use amaru_kernel::{Nullable, Tx, TxId};
+    use amaru_kernel::{Nullable, Tx};
     use amaru_ouroboros_traits::{
-        CanValidateTransactions, MempoolSeqNo, TransactionValidationError, TxOrigin,
+        CanValidateTransactions, MempoolSeqNo, TransactionValidationError, TxId, TxOrigin,
         TxRejectReason, TxSubmissionMempool,
     };
     use pallas_primitives::Set;
@@ -355,7 +354,7 @@ mod tests {
             tx: Tx,
             _tx_origin: TxOrigin,
         ) -> Result<(TxId, MempoolSeqNo), TxRejectReason> {
-            Ok((TxId::from(tx), MempoolSeqNo(1)))
+            Ok((TxId::from(&tx), MempoolSeqNo(1)))
         }
 
         fn get_tx(&self, _tx_id: &TxId) -> Option<Arc<Tx>> {
@@ -367,7 +366,7 @@ mod tests {
             _from_seq: MempoolSeqNo,
             _limit: u16,
         ) -> Vec<(TxId, u32, MempoolSeqNo)> {
-            vec![(TxId::from(self.tx.clone()), 100, MempoolSeqNo(1))]
+            vec![(TxId::from(&self.tx), 100, MempoolSeqNo(1))]
         }
 
         fn wait_for_at_least(
