+++ crates/amaru-consensus/src/consensus/mod.rs
@@ -18,5 +18,6 @@ pub mod headers_tree;
 pub mod span;
 pub mod stages;
 pub mod store;
+pub mod tx_submission;
 
 pub const EVENT_TARGET: &str = "amaru::consensus";
