+++ crates/amaru-consensus/src/consensus/tx_submission/mod.rs
@@ -12,26 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod has_stake_distribution;
-pub use has_stake_distribution::{HasStakeDistribution, PoolSummary};
+pub mod server_params;
+pub mod tx_submission_client_state;
+pub mod tx_submission_server_state;
 
-pub mod can_validate_blocks;
-pub use can_validate_blocks::{BlockValidationError, CanValidateBlocks};
+pub use server_params::*;
+pub use tx_submission_client_state::TxSubmissionClientState;
+pub use tx_submission_server_state::TxSubmissionServerState;
 
-pub mod can_validate_transactions;
-pub use can_validate_transactions::{CanValidateTransactions, TransactionValidationError};
-
-pub mod stores;
-pub use stores::*;
-
-pub mod praos;
-pub use praos::*;
-
-pub mod mempool;
-pub use mempool::*;
-
-pub mod network_operations;
-pub use network_operations::NetworkOperations;
-
-pub mod tx_submission_events;
-pub use tx_submission_events::*;
+#[cfg(test)]
+mod tests;
