+++ crates/amaru-consensus/src/consensus/tx_submission/tests/sized_mempool.rs
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::tx_submission::tests::Tx;
-use amaru_kernel::tx_submission_events::TxId;
 use amaru_mempool::strategies::InMemoryMempool;
 use amaru_mempool::{DefaultCanValidateTransactions, MempoolConfig};
 use amaru_ouroboros_traits::{
-    CanValidateTransactions, Mempool, MempoolSeqNo, TransactionValidationError, TxOrigin,
+    CanValidateTransactions, Mempool, MempoolSeqNo, TransactionValidationError, TxId, TxOrigin,
     TxRejectReason, TxSubmissionMempool,
 };
+use pallas_primitives::conway::Tx;
 use std::pin::Pin;
 use std::sync::Arc;
