+++ crates/amaru-consensus/src/consensus/tx_submission/tests/test_cases.rs
@@ -12,12 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::tx_submission::tests::{
-    FaultyTxValidator, ServerOptions, Tx, create_node, create_node_with, expect_server_transactions,
+use crate::consensus::tx_submission::Blocking;
+use crate::consensus::tx_submission::tests::{
+    FaultyTxValidator, ServerOptions, create_node, create_node_with, expect_server_transactions,
 };
-use crate::tx_submission::{Blocking, new_era_tx_id};
+use amaru_kernel::{Nullable, Set};
+use amaru_network::era_tx_ids;
+use amaru_ouroboros_traits::TxId;
+use pallas_crypto::hash::Hash;
 use pallas_network::miniprotocols::txsubmission::Message::{RequestTxIds, RequestTxs};
 use pallas_network::miniprotocols::txsubmission::{EraTxBody, EraTxId, Message};
+use pallas_primitives::TransactionInput;
+use pallas_primitives::conway::{PseudoTransactionBody, PseudoTx, Tx, WitnessSet};
 use std::sync::Arc;
 
 #[tokio::test]
@@ -73,7 +79,8 @@ async fn test_client_with_non_blocking_server() -> anyhow::Result<()> {
 #[tokio::test]
 async fn test_invalid_transactions() -> anyhow::Result<()> {
     let txs = create_transactions(6);
-    let era_tx_ids: Vec<EraTxId> = txs.iter().map(|tx| new_era_tx_id(tx.tx_id())).collect();
+    let tx_ids: Vec<TxId> = txs.iter().map(TxId::from).collect();
+    let era_tx_ids = era_tx_ids(&tx_ids);
 
     // This validator rejects every second transaction
     let tx_validator = Arc::new(FaultyTxValidator::default());
@@ -114,10 +121,55 @@ async fn test_invalid_transactions() -> anyhow::Result<()> {
 
 // HELPERS
 
-pub fn create_transactions(number: u16) -> Vec<Tx> {
+pub fn create_transactions(number: u64) -> Vec<Tx> {
     let mut txs = vec![];
     for i in 0..number {
-        txs.push(Tx::new(format!("{i}d8d00cdd4657ac84d82f0a56067634a")));
+        txs.push(create_transaction(i));
     }
     txs
+}
+
+pub fn create_transaction(number: u64) -> Tx {
+    let tx_input = TransactionInput {
+        transaction_id: Hash::new([1; 32]),
+        index: number,
+    };
+    let transaction_body = PseudoTransactionBody {
+        inputs: Set::from(vec![tx_input]),
+        outputs: vec![],
+        fee: 0,
+        ttl: None,
+        certificates: None,
+        withdrawals: None,
+        auxiliary_data_hash: None,
+        validity_interval_start: None,
+        mint: None,
+        script_data_hash: None,
+        required_signers: None,
+        network_id: None,
+        collateral_return: None,
+        total_collateral: None,
+        reference_inputs: None,
+        voting_procedures: None,
+        proposal_procedures: None,
+        treasury_value: None,
+        collateral: None,
+        donation: None,
+    };
+    let transaction_witness_set = WitnessSet {
+        vkeywitness: None,
+        native_script: None,
+        bootstrap_witness: None,
+        plutus_v1_script: None,
+        plutus_data: None,
+        redeemer: None,
+        plutus_v2_script: None,
+        plutus_v3_script: None,
+    };
+    PseudoTx {
+        transaction_body,
+        transaction_witness_set,
+        success: true,
+        auxiliary_data: Nullable::Null,
+    }
 }
