+++ crates/amaru-network/src/tx_submission/mod.rs
@@ -12,17 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod conversions;
-mod server_params;
-mod tx_client_transport;
-mod tx_server_transport;
-pub mod tx_submission_client;
-pub mod tx_submission_server;
+pub mod conversions;
 
 pub use conversions::*;
-pub use server_params::*;
-pub use tx_submission_client::TxSubmissionClientState;
-pub use tx_submission_server::TxSubmissionServerState;
-
-#[cfg(test)]
-mod tests;
