+++ simulation/amaru-sim/src/simulator/mock_network_resource.rs
@@ -20,7 +20,8 @@ use amaru_consensus::network_operations::ForwardEvent;
 use amaru_kernel::connection::ClientConnectionError;
 use amaru_kernel::consensus_events::{ChainSyncEvent, Tracked};
 use amaru_kernel::peer::Peer;
-use amaru_kernel::{IsHeader, Point, TxClientReply, TxServerRequest, to_cbor};
+use amaru_kernel::{IsHeader, Point, to_cbor};
+use amaru_ouroboros::{TxClientReply, TxServerRequest};
 use async_trait::async_trait;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
